Update README.md
Introduction:
A number guessing game is a simple and fun way to learn about programming and the Java programming language. In this article, we will walk through the process of creating a number guessing game using Java.

Explanation:
The first step in creating a number guessing game is to generate a random number for the player to guess. In Java, this can be done using the Random class. The nextInt() method of this class generates a random integer between 0 and the specified maximum value. For our game, we want the number to be between 1 and 100, so we will use the following code to generate our random number: